+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 6.0.0a110 (unreleased)
 ----------------------
 
-- Nothing changed yet.
-
+- Integration with redturtle.prenotazioni
+  [foxtrot-dfm1]
 
 6.0.0a19 (2023-02-27)
 ---------------------